refactor : position 패키지 이동에 따른 변경
