+++ src/main/java/janggi/domain/board/Board.java
@@ -3,7 +3,7 @@ package janggi.domain.board;
 import janggi.domain.piece.EmptyPiece;
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.Team;
-import janggi.domain.vo.Position;
+import janggi.domain.vo.position.Position;
 
 import java.util.HashMap;
 import java.util.Map;
