+++ src/test/java/janggi/domain/DirectionTest.java
@@ -1,6 +1,6 @@
 package janggi.domain;
 
-import janggi.domain.vo.Position;
+import janggi.domain.vo.position.Position;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
