+++ src/test/java/janggi/domain/board/BoardTest.java
@@ -1,8 +1,7 @@
 package janggi.domain.board;
 
-import janggi.domain.FakeBoard;
 import janggi.domain.piece.*;
-import janggi.domain.vo.Position;
+import janggi.domain.vo.position.Position;
 import org.junit.jupiter.api.Test;
 
 import java.util.HashMap;
