+++ src/test/java/janggi/domain/piece/ElephantTest.java
@@ -2,7 +2,7 @@ package janggi.domain.piece;
 
 import janggi.domain.FakeBoard;
 import janggi.domain.board.BoardView;
-import janggi.domain.vo.Position;
+import janggi.domain.vo.position.Position;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
