+++ src/test/java/janggi/domain/vo/PositionTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import janggi.domain.Direction;
-import org.junit.jupiter.api.DisplayName;
+import janggi.domain.vo.position.Position;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
